Add KMIP support to getkey
This change adds KMIP backend support to the getkey utility,
allowing keys to be fetched from a KMIP server given a valid
TLs connection to the server and the ID of the key to retrieve.
Communications with the KMIP server are managed by libkmip,
which encodes and decodes the KMIP request and response
messages that are sent across the BIO connection.

KMIP support is implemented as another utility function, leaving
room for a CLI flag for getkey that can toggle which type of
backend to use.

The project Makefile is also updated to link libkmip during the
build process.