Add error checking to tpm2_unpack_uint32_to_str (#40)
This change updates the tpm2_unpack_uint32_to_str utility function
to support error checking the asprintf call. The function now
returns an int status code for success or failure and logs an error
when it occurs. The usage of this function in tpm2_get_impl_type
has been updated to also handle the new error... handling.

Partially addresses #25