Updating the function signature for get_ski_block_bytes
This change updates the function signature for get_ski_block_bytes,
now requiring the lengths of the current and next expected
delimeters to be passed along with the strings themselves. This
simplifies the internal get_ski_block_bytes code and sets up a
future change using strnlen for string length.

Partially addresses #25