Update the Kmyth SGX Makefile to install libraries
This change updates the Kmyth SGX Makefile, adding install and
uninstall sections for the Kmyth SGX libraries and headers that
place them alongside the existing Kmyth libraries and headers.

The install location of the signed seal library is now passed to
the build process as a preprocessor directive, which is detected
and used as the enclave path by the underlying source code.

Finally, the names of the Kmyth SGX libraries have been updated.

Closes #52