Add an initial ECDH key server for testing purposes
This change adds a basic implementation of an ECDH-based key server
for use with the SGX testing infrastructure. The server currently
loads in its signing key and the public key of the client it will
talk to before dynamically generating an ephemeral key to be used
to establish a shared session key. Placeholders indicate where
message signing and verification must occur.

NOTE: this server is still undergoing active test and development
and will need additional updates before it's fully operational.
This commit will likely be clobbered with an update in the near
future.