Fix KMIP key id length bug.
It is not necessary to include a terminating null byte
or account for it in the id length.

The non-SGX client also had an incorrect length value
because it used sizeof after casting the string pointer.

Also change the hard-coded key id to a realistic value
that can be used with the PyKMIP test server.